cilium: Move common/config.go to pkg/option/config.go - fix #760
